+++ R/tibble-package.R
@@ -24,7 +24,7 @@
 #'
 #' Resources on specific topics:
 #'   * Create a tibble: [tibble()], [as_tibble()], [tribble()], [enframe()]
-#'   * Inspect a tibble: [print.tbl_df()], [glimpse()]
+#'   * Inspect a tibble: [print.tbl()], [glimpse()]
 #'   * Details on the S3 `tbl_df` class: [`tbl_df-class`]
 "_PACKAGE"