+++ R/utils-format.r
@@ -17,7 +17,7 @@
 #'
 #' Options used by the tibble and pillar packages to format and print `tbl_df`
 #' objects. Used by the formatting workhorse `trunc_mat()` and, therefore,
-#' indirectly, by `print.tbl_df()` and `print.tbl()`.
+#' indirectly, by `print.tbl()`.
 #'
 #' * `tibble.print_max`: Row number threshold: Maximum number of rows printed.
 #'   Set to `Inf` to always print all rows.  Default: 20.
@@ -60,6 +60,20 @@
 #' }
 #' @name formatting
 NULL
+
+#' @rdname formatting
+#' @export
+print.tbl <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
+  cat_line(format(x, ..., n = n, width = width, n_extra = n_extra))
+  invisible(x)
+}
+
+#' @rdname formatting
+#' @export
+format.tbl <- function(x, ..., n = NULL, width = NULL, n_extra = NULL) {
+  mat <- trunc_mat(x, n = n, width = width, n_extra = n_extra)
+  format(mat)
+}
 
 #' @export
 #' @rdname formatting